backend/requirements.txt
# AI Floor Plan and Market Insights - Backend Dependencies
# REQUIRES: Python 3.10+ (CrewAI dependency)
# RECOMMENDED: Python 3.11+

# ================================
# Core Framework
# ================================
Flask==3.0.3
Flask-CORS==4.0.0
Werkzeug==3.0.3

# ================================
# Authentication & Security
# ================================
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
cryptography==42.0.5

# ================================
# Database & ORM
# ================================
supabase==2.4.4
psycopg2-binary==2.9.9

# ================================
# Async Task Queue
# ================================
celery==5.3.6
redis==5.0.3

# ================================
# AI & LLM (Phase 2+)
# ================================
google-generativeai==0.4.1
crewai==0.1.24  # Pinned to work with langchain 0.0.354
crewai-tools>=0.1.0
langchain==0.0.354
langchain-google-genai==0.0.11
litellm>=1.40.0

# ================================
# API Clients
# ================================
requests==2.31.0
httpx==0.27.0
tavily-python

# ================================
# Web Scraping (Phase 6)
# ================================
playwright==1.40.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ================================
# Image Processing & OCR
# ================================
Pillow==10.3.0
pytesseract==0.3.10  # OCR fallback for dimension extraction

# ================================
# Data Validation
# ================================
pydantic>=2.7.4

# ================================
# Testing
# ================================
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-flask==1.3.0
pytest-asyncio==0.21.1

# ================================
# Statistical Analysis (Phase 6)
# ================================
scikit-learn==1.5.2
pandas==2.2.1
numpy==1.26.4

# ================================
# Utilities
# ================================
python-dateutil==2.9.0

# NOTE: AI dependencies (CrewAI, LangChain) will be added in Phase 1
# to avoid dependency conflicts during initial setup
